Implement GraphQL endpoints and DB setup
Implemented CreatePost, UpdatePost, DeletePost, Post, and Posts resolvers in graph/schema.resolvers.go. On top of that, database connection was established and migrated with the Post model in the new database/db.go file. Package github.com/joho/godotenv is used to load environment variables for DB setup from a .env file.

Reason:
This update was necessary to set up necessary GraphQL endpoints for CRUD operations and establish a connection with the Postgres database. The environment variables, which configure the database connection, are now loaded from a .env file which improves maintainability and security.